Finished implementation of GetUpdate controller method
The database page no longer reloads the entire list when there is a change in the database, instead it passes its current list of persons to the GetUpdate method of the database controller.

The GetUpdate method loads the Person list from the viewer and the database and compares them to see if there is a change. It returns two JSON objects, one contains a list of Persons that need to be changed or added to the viewer and the other contains a list of Persons that need to be deleted from the viewer.